Solve problem 4, finding the largest palindrome with 3-digit factors
Reworked palindrome generation to yield palindromes in descending order,
and added a function to check whether a number was a factor of two
n-digit numbers. Since the number of palindromes is relatively small,
it's enough to just iterate through in descenting order and pick the
first match.

Also re-worked Highest Prime Factor to be expressed in terms of finding
all the prime factors, since that's what's required anyway.