challenge: part 8.
Added functionality to PrimeBag to handle comparisons between large (factored) numbers.
Added multiplication and division on PrimeBag.
Added rudimentary ordering between PrimeBags.

Next: change the backing structure of PrimeBag to properly handle division by another PrimeBag which is not a factor of the first (currently just ignores factors in `other` that are not present in `self`)